nodes: move livectls API to scene
These functions already work with a scene as input. We may want to
rename these functions to use "ngl_scene_" as prefix at some point.
