+++ libnopegl/src/api.c
@@ -795,16 +795,6 @@ int ngl_gl_wrap_framebuffer(struct ngl_ctx *s, uint32_t framebuffer)
     return 0;
  }
 
-int ngl_livectls_get(struct ngl_scene *scene, size_t *nb_livectlsp, struct ngl_livectl **livectlsp)
-{
-    return ngli_node_livectls_get(scene, nb_livectlsp, livectlsp);
-}
-
-void ngl_livectls_freep(struct ngl_livectl **livectlsp)
-{
-    ngli_node_livectls_freep(livectlsp);
-}
-
 void ngl_freep(struct ngl_ctx **ss)
 {
     struct ngl_ctx *s = *ss;
